requirements.txt
fastapi==0.96.0
pandas==2.0.2
Pillow==9.5.0
scikit-learn==1.2.2
tensorflow==2.12.0
uvicorn==0.22.0
openpyxl
python-multipart